Removed HintButton.jsx component as it is not working with js, dont know why.
Convert the HintButton.jsx to GenerateHintButton.js
